Test upgrade (but not downgrade) in all sanity tests*
* Except template
