+++ frameworks/cassandra/tests/test_backup_and_restore.py
@@ -19,7 +19,7 @@ VERIFY_DELETION_JOB = get_verify_deletion_job(node_address=FOLDERED_NODE_ADDRESS
 TEST_JOBS = [WRITE_DATA_JOB, VERIFY_DATA_JOB, DELETE_DATA_JOB, VERIFY_DELETION_JOB]
 
 no_strict = pytest.mark.skipif(os.environ.get("SECURITY") == "strict",
-        reason="upgrade/downgrade tests broken in strict")
+        reason="backup/restore tests broken in strict")
 
 @pytest.fixture(scope='module', autouse=True)
 def configure_package(configure_universe):
@@ -27,15 +27,12 @@ def configure_package(configure_universe):
         sdk_install.uninstall(FOLDERED_SERVICE_NAME, package_name=PACKAGE_NAME)
         sdk_utils.gc_frameworks()
 
-        # 1. check_suppression=False due to https://jira.mesosphere.com/browse/CASSANDRA-568
-        # 2. user: root because Azure CLI needs to run in root...
+        # user=root because Azure CLI needs to run in root...
         sdk_install.install(
             PACKAGE_NAME,
             DEFAULT_TASK_COUNT,
             service_name=FOLDERED_SERVICE_NAME,
-            additional_options={"service": { "name": FOLDERED_SERVICE_NAME, "user": "root" } },
-            check_suppression=False)
-        sdk_plan.wait_for_completed_deployment(FOLDERED_SERVICE_NAME)
+            additional_options={"service": { "name": FOLDERED_SERVICE_NAME, "user": "root" } })
 
         tmp_dir = tempfile.mkdtemp(prefix='cassandra-test')
         for job in TEST_JOBS:
