+++ frameworks/cassandra/tests/test_sanity.py
@@ -9,9 +9,10 @@ import sdk_cmd as cmd
 import sdk_hosts
 import sdk_install
 import sdk_jobs
-import sdk_plan
-import sdk_utils
 import sdk_metrics
+import sdk_plan
+import sdk_upgrade
+import sdk_utils
 
 
 WRITE_DATA_JOB = get_write_data_job(node_address=FOLDERED_NODE_ADDRESS)
@@ -28,14 +29,12 @@ def configure_package(configure_universe):
         sdk_install.uninstall(FOLDERED_SERVICE_NAME, package_name=PACKAGE_NAME)
         sdk_utils.gc_frameworks()
 
-        # check_suppression=False due to https://jira.mesosphere.com/browse/CASSANDRA-568
-        sdk_install.install(
+        sdk_upgrade.test_upgrade(
+            "beta-{}".format(PACKAGE_NAME),
             PACKAGE_NAME,
             DEFAULT_TASK_COUNT,
             service_name=FOLDERED_SERVICE_NAME,
-            additional_options={"service": { "name": FOLDERED_SERVICE_NAME } },
-            check_suppression=False)
-        sdk_plan.wait_for_completed_deployment(FOLDERED_SERVICE_NAME)
+            additional_options={"service": {"name": FOLDERED_SERVICE_NAME} })
 
         tmp_dir = tempfile.mkdtemp(prefix='cassandra-test')
         for job in TEST_JOBS:
