+++ frameworks/cassandra/tests/test_soak.py
@@ -11,7 +11,7 @@ import sdk_api
 import sdk_hosts
 import sdk_jobs
 import sdk_plan
-import sdk_test_upgrade
+import sdk_upgrade
 import sdk_utils
 
 
@@ -47,10 +47,12 @@ def test_soak_upgrade_downgrade():
     Assumes that the install options file is placed in the repo root."""
     with open('cassandra.json') as options_file:
         install_options = json.load(options_file)
-
-    sdk_test_upgrade.soak_upgrade_downgrade(
-        PACKAGE_NAME, DEFAULT_TASK_COUNT, install_options
-    )
+    sdk_upgrade.soak_upgrade_downgrade(
+        "beta-{}".format(PACKAGE_NAME),
+        PACKAGE_NAME,
+        DEFAULT_TASK_COUNT,
+        service_name=install_options["service"]["name"],
+        additional_options=install_options)
 
 
 @pytest.mark.soak_migration
