+++ frameworks/elastic/tests/test_shakedown.py
@@ -8,7 +8,7 @@ import sdk_install
 import sdk_marathon
 import sdk_metrics
 import sdk_tasks
-import sdk_test_upgrade
+import sdk_upgrade
 import sdk_utils
 from tests.config import *
 
@@ -19,11 +19,13 @@ def configure_package(configure_universe):
     try:
         sdk_install.uninstall(FOLDERED_SERVICE_NAME, package_name=PACKAGE_NAME)
         sdk_utils.gc_frameworks()
-        sdk_install.install(
+
+        sdk_upgrade.test_upgrade(
+            "beta-{}".format(PACKAGE_NAME),
             PACKAGE_NAME,
             DEFAULT_TASK_COUNT,
             service_name=FOLDERED_SERVICE_NAME,
-            additional_options={"service": { "name": FOLDERED_SERVICE_NAME } })
+            additional_options={"service": {"name": FOLDERED_SERVICE_NAME}})
 
         yield # let the test session execute
     finally:
