+++ frameworks/hdfs/tests/test_shakedown.py
@@ -8,10 +8,11 @@ import sdk_cmd as cmd
 import sdk_hosts
 import sdk_install
 import sdk_marathon
+import sdk_metrics
 import sdk_plan
 import sdk_tasks
+import sdk_upgrade
 import sdk_utils
-import sdk_metrics
 from tests.config import *
 
 
@@ -20,12 +21,13 @@ def configure_package(configure_universe):
     try:
         sdk_install.uninstall(FOLDERED_SERVICE_NAME, package_name=PACKAGE_NAME)
         sdk_utils.gc_frameworks()
-        sdk_install.install(
+
+        sdk_upgrade.test_upgrade(
+            "beta-{}".format(PACKAGE_NAME),
             PACKAGE_NAME,
             DEFAULT_TASK_COUNT,
             service_name=FOLDERED_SERVICE_NAME,
-            additional_options={"service": { "name": FOLDERED_SERVICE_NAME } })
-        sdk_plan.wait_for_completed_deployment(FOLDERED_SERVICE_NAME)
+            additional_options={"service": {"name": FOLDERED_SERVICE_NAME}})
 
         yield # let the test session execute
     finally:
