+++ frameworks/helloworld/tests/test_secrets.py
@@ -91,10 +91,6 @@ def test_secrets_basic():
 
     sdk_install.install(PACKAGE_NAME, NUM_HELLO + NUM_WORLD, additional_options=secret_options)
 
-    # default is serial strategy, hello deploys first
-    # launch will fail if secrets are not available or not accessible
-    sdk_plan.wait_for_completed_deployment(PACKAGE_NAME)
-
     hello_tasks_0 = sdk_tasks.get_task_ids(PACKAGE_NAME, "hello-0")
     world_tasks_0 = sdk_tasks.get_task_ids(PACKAGE_NAME, "word-0")
 
@@ -127,9 +123,6 @@ def test_secrets_verify():
     create_secrets("{}/".format(PACKAGE_NAME))
 
     sdk_install.install(PACKAGE_NAME, NUM_HELLO + NUM_WORLD, additional_options=secret_options)
-
-    # launch will fail if secrets are not available or not accessible
-    sdk_plan.wait_for_completed_deployment(PACKAGE_NAME)
 
     # tasks will fail if secret file is not created
     sdk_tasks.check_running(PACKAGE_NAME, NUM_HELLO + NUM_WORLD)
@@ -184,9 +177,6 @@ def test_secrets_update():
     create_secrets("{}/".format(PACKAGE_NAME))
 
     sdk_install.install(PACKAGE_NAME, NUM_HELLO + NUM_WORLD, additional_options=secret_options)
-
-    # launch will fail if secrets are not available or not accessible
-    sdk_plan.wait_for_completed_deployment(PACKAGE_NAME)
 
     # tasks will fail if secret file is not created
     sdk_tasks.check_running(PACKAGE_NAME, NUM_HELLO + NUM_WORLD)
@@ -244,9 +234,6 @@ def test_secrets_config_update():
     create_secrets("{}/".format(PACKAGE_NAME))
 
     sdk_install.install(PACKAGE_NAME, NUM_HELLO + NUM_WORLD, additional_options=secret_options)
-
-    # launch will fail if secrets are not available or not accessible
-    sdk_plan.wait_for_completed_deployment(PACKAGE_NAME)
 
     # tasks will fail if secret file is not created
     sdk_tasks.check_running(PACKAGE_NAME, NUM_HELLO + NUM_WORLD)
