+++ frameworks/helloworld/tests/test_soak.py
@@ -9,7 +9,7 @@ import sdk_install
 import sdk_plan
 import sdk_tasks
 import sdk_marathon
-import sdk_test_upgrade
+import sdk_upgrade
 from tests.config import (
     PACKAGE_NAME,
     DEFAULT_TASK_COUNT
@@ -30,7 +30,7 @@ if "NUM_WORLD" in os.environ:
 
 @pytest.mark.soak_upgrade
 def test_soak_upgrade_downgrade():
-    sdk_test_upgrade.soak_upgrade_downgrade(PACKAGE_NAME, PACKAGE_NAME, DEFAULT_TASK_COUNT)
+    sdk_upgrade.soak_upgrade_downgrade(PACKAGE_NAME, PACKAGE_NAME, PACKAGE_NAME, DEFAULT_TASK_COUNT)
 
 
 @pytest.mark.soak_secrets_update
