+++ frameworks/kafka/tests/test_sanity.py
@@ -4,11 +4,12 @@ import urllib
 import sdk_hosts
 import sdk_install as install
 import sdk_marathon
+import sdk_metrics
 import sdk_plan
 import sdk_tasks
+import sdk_upgrade
 import sdk_utils
-import sdk_metrics
-import shakedown
+
 import dcos
 import dcos.config
 import dcos.http
@@ -28,12 +29,13 @@ def configure_package(configure_universe):
     try:
         install.uninstall(FOLDERED_SERVICE_NAME, package_name=PACKAGE_NAME)
         sdk_utils.gc_frameworks()
-        install.install(
+
+        sdk_upgrade.test_upgrade(
+            "beta-{}".format(PACKAGE_NAME),
             PACKAGE_NAME,
             DEFAULT_BROKER_COUNT,
             service_name=FOLDERED_SERVICE_NAME,
-            additional_options={"service": { "name": FOLDERED_SERVICE_NAME } })
-        sdk_plan.wait_for_completed_deployment(FOLDERED_SERVICE_NAME)
+            additional_options={"service": {"name": FOLDERED_SERVICE_NAME}})
 
         yield # let the test session execute
     finally:
